create AWSReportDataGenerator. move aws tests into aws folder.
